fix: align to changes of python gen structure
