+++ python/olink_tests/context.py
@@ -1,0 +1,4 @@
+# import os
+# import sys
+# sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../test_api')))
+
